Refactored auth modal common logic into seperate BaseModal
