+++ packages/commonwealth/client/scripts/views/modals/AuthModal/BaseModal/BaseModal.scss
@@ -1,6 +1,6 @@
-@import '../../../../styles/shared';
+@import '../../../../../styles/shared';
 
-.AuthModal {
+.BaseModal {
   display: flex;
   flex-direction: column;
   position: relative;
