+++ packages/commonwealth/client/scripts/views/modals/AuthModal/BaseModal/index.ts
@@ -1,0 +1,1 @@
+export { BaseModal } from './BaseModal';
