+++ packages/commonwealth/client/scripts/views/modals/AuthModal/types.ts
@@ -6,8 +6,7 @@ export type AuthModalTabs = {
   options: AuthWallets[] | AuthSSOs[];
 };
 
-export type AuthModalProps = {
-  isOpen: boolean;
+export type BaseModalProps = {
   onClose: () => any;
   onSuccess?: () => any;
   showWalletsFor?:
@@ -16,3 +15,7 @@ export type AuthModalProps = {
     | ChainBase.Solana
     | ChainBase.Substrate;
 };
+
+export type AuthModalProps = {
+  isOpen: boolean;
+} & BaseModalProps;
